feat: 입력값 맞추어 모델 수정 #14
