+++ app/main.py
@@ -1,9 +1,7 @@
-from fastapi import FastAPI, Header, HTTPException
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from typing import Optional
-from app.routers import health, chat, news
-from app.models.chat import ChatRequest, ChatResponse
-from app.services.rag_service import process_chat
+
+from app.routers import health, chat, news, coin_predict
 
 app = FastAPI(title="EyeStock Chat API", version="1.0.0")
 
@@ -18,3 +16,4 @@ app.add_middleware(
 app.include_router(health.router)
 app.include_router(chat.router)
 app.include_router(news.router)
+app.include_router(coin_predict.router)
